test(llc): migrate channel location API tests to stream_chat_test
