+++ packages/stream_chat/test/src/client/channel/channel_test.dart
@@ -1074,109 +1074,6 @@ void main() {
       );
     });
 
-    group('`.sendStaticLocation`', () {
-      const deviceId = 'test-device-id';
-      const locationId = 'test-location-id';
-      const coordinates = LocationCoordinates(
-        latitude: 40.7128,
-        longitude: -74.0060,
-      );
-
-      test('should create a static location and call sendMessage', () async {
-        when(
-          () => client.sendMessage(any(), channelId, channelType),
-        ).thenAnswer(
-          (_) async => SendMessageResponse()
-            ..message = Message(
-              id: locationId,
-              text: 'Location shared',
-              extraData: const {'custom': 'data'},
-              sharedLocation: Location(
-                channelCid: channel.cid,
-                messageId: locationId,
-                userId: client.state.currentUser?.id,
-                latitude: coordinates.latitude,
-                longitude: coordinates.longitude,
-                createdByDeviceId: deviceId,
-              ),
-            ),
-        );
-
-        final response = await channel.sendStaticLocation(
-          id: locationId,
-          messageText: 'Location shared',
-          createdByDeviceId: deviceId,
-          location: coordinates,
-          extraData: {'custom': 'data'},
-        );
-
-        expect(response, isNotNull);
-        expect(response.message.id, locationId);
-        expect(response.message.text, 'Location shared');
-        expect(response.message.extraData['custom'], 'data');
-        expect(response.message.sharedLocation, isNotNull);
-
-        verify(
-          () => client.sendMessage(any(), channelId, channelType),
-        ).called(1);
-      });
-    });
-
-    group('`.startLiveLocationSharing`', () {
-      const deviceId = 'test-device-id';
-      const locationId = 'test-location-id';
-      final endSharingAt = DateTime.timestamp().add(const Duration(hours: 1));
-      const coordinates = LocationCoordinates(
-        latitude: 40.7128,
-        longitude: -74.0060,
-      );
-
-      test(
-        'should create message with live location and call sendMessage',
-        () async {
-          when(
-            () => client.sendMessage(any(), channelId, channelType),
-          ).thenAnswer(
-            (_) async => SendMessageResponse()
-              ..message = Message(
-                id: locationId,
-                text: 'Location shared',
-                extraData: const {'custom': 'data'},
-                sharedLocation: Location(
-                  channelCid: channel.cid,
-                  messageId: locationId,
-                  userId: client.state.currentUser?.id,
-                  latitude: coordinates.latitude,
-                  longitude: coordinates.longitude,
-                  createdByDeviceId: deviceId,
-                  endAt: endSharingAt,
-                ),
-              ),
-          );
-
-          final response = await channel.startLiveLocationSharing(
-            id: locationId,
-            messageText: 'Location shared',
-            createdByDeviceId: deviceId,
-            location: coordinates,
-            endSharingAt: endSharingAt,
-            extraData: {'custom': 'data'},
-          );
-
-          expect(response, isNotNull);
-          expect(response.message.id, locationId);
-          expect(response.message.text, 'Location shared');
-          expect(response.message.extraData['custom'], 'data');
-          expect(response.message.sharedLocation, isNotNull);
-          expect(response.message.sharedLocation?.endAt, endSharingAt);
-
-          verify(
-            () => client.sendMessage(any(), channelId, channelType),
-          ).called(1);
-        },
-      );
-    });
-
     group('`.sendReaction`', () {
       test('should work fine', () async {
         final message = Message(
